Implemented MIDITriggerable
