+++ Sources/AudioKit/Nodes/Generators/Physical Models/Clarinet.swift
@@ -4,7 +4,7 @@ import AVFoundation
 
 /// STK Clarinet
 ///
-public class Clarinet: Node {
+public class Clarinet: Node, MIDITriggerable {
 
     /// Connected nodes
     public var connections: [Node] { [] }
@@ -14,15 +14,4 @@ public class Clarinet: Node {
     
     /// Initialize the STK Clarinet model
     public init() {}
-
-    /// Trigger the sound with a set of parameters
-    ///
-    /// - Parameters:
-    ///   - note: MIDI note number
-    ///   - velocity: Amplitude or volume expressed as a MIDI Velocity 0-127
-    ///
-    public func trigger(note: MIDINoteNumber, velocity: MIDIVelocity = 127) {
-        start()
-        auBase.trigger(note: note, velocity: velocity)
-    }
 }
