+++ Sources/AudioKit/Nodes/Generators/Physical Models/Flute.swift
@@ -4,7 +4,7 @@ import AVFoundation
 
 /// STK Flute
 ///
-public class Flute: Node {
+public class Flute: Node, MIDITriggerable {
     
     /// Connected nodes
     public var connections: [Node] { [] }
@@ -14,15 +14,4 @@ public class Flute: Node {
 
     /// Initialize the STK Flute model
     public init() {}
-
-    /// Trigger the sound with a set of parameters
-    ///
-    /// - Parameters:
-    ///   - note: MIDI note number
-    ///   - velocity: Amplitude or volume expressed as a MIDI Velocity 0-127
-    ///
-    public func trigger(note: MIDINoteNumber, velocity: MIDIVelocity = 127) {
-        start()
-        auBase.trigger(note: note, velocity: velocity)
-    }
 }
