+++ Sources/AudioKit/Nodes/Generators/Physical Models/TubularBells.swift
@@ -4,7 +4,7 @@ import AVFoundation
 
 /// STK Tubuluar Bells
 ///
-public class TubularBells: Node {
+public class TubularBells: Node, MIDITriggerable {
     
     /// Connected nodes
     public var connections: [Node] { [] }
@@ -14,15 +14,4 @@ public class TubularBells: Node {
 
     /// Initialize the STK Tubular Bells model
     public init() { }
-
-    /// Trigger the sound with a set of parameters
-    ///
-    /// - Parameters:
-    ///   - note: MIDI note number
-    ///   - velocity: Amplitude or volume expressed as a MIDI Velocity 0-127
-    ///
-    public func trigger(note: MIDINoteNumber, velocity: MIDIVelocity = 127) {
-        start()
-        auBase.trigger(note: note, velocity: velocity)
-    }
 }
