+++ Sources/AudioKit/Nodes/Node.swift
@@ -172,19 +172,6 @@ extension Node {
     }
 }
 
-/// To allow physical models to be triggered
-public protocol Triggerable {
-    func trigger()
-}
-
-extension Node where Self: Triggerable {
-    /// Trigger the sound with current parameters
-    ///
-    public func trigger() {
-        auBase.trigger()
-    }
-}
-
 func instantiate(componentDescription: AudioComponentDescription) -> AVAudioUnit {
 
     let semaphore = DispatchSemaphore(value: 0)
